+++ index.html
@@ -7,7 +7,7 @@
   <title>Simple Platform Game</title>
 </head>
 <body>
-  <script src="productionDependencies.bundle.js"></script>
+  <script src="loader.bundle.js"></script>
   <script src="app.bundle.js"></script>
 </body>
 </html>
